fix(Recette): Correction d'un bug ne permettant pas de sauvegarder les ingredients et les etapes lors du save d'une recette

+++ api/src/main/java/com/api/kookie/controllers/recette/RecetteController.java
@@ -36,7 +36,7 @@ public class RecetteController {
     }
 
     @GetMapping("/one")
-    public ResponseEntity<?> getOneByRecipeId(@RequestParam Long recipeId) {
+    public ResponseEntity<?> getOneByRecipeId(@RequestParam Integer recipeId) {
         LOGGER.debug("[RecetteController, getOneByRecipeId] recipeId = " + recipeId);
 
         RecetteDTO recipe = recetteService.getOneById(recipeId);
